reduce qss-bind
