+++ releasenotes/notes/refactor_bind-ab3c07883604767b.yaml
@@ -1,8 +1,4 @@
 ---
-features:
-  - |
-    Added a new command ``qss-bind`` that has the same capability of ``link_file``
-    in python_lib.
 upgrade:
   - |
     Allow target-specific parameter bindings when ``qe-qem`` is specified as
