--- tools/qss-bind/qss-bind.cpp
@@ -1,5 +1,0 @@
-#include "API/api.h"
-
-int main(int argc, const char **argv) {
-  return qssc::bind(argc, argv, nullptr, {});
-}
